client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12,<3.14"
dependencies = [
    "pywebview (>=6.1,<7.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "websockets (>=16.0,<17.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "kyber-py (>=1.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.2,<0.16.0)"
]
